JS Update entry point path
Hey there Flipside gang! I was working on a little side project that uses ShroomDK and ran into some issues with importing the module. 

In my node modules, the index of the package is found in @flipsidecrypto/sdk/dist/src/index.js, but the base package.json has the main and type entry points as just dist/index.js. Making the change I've proposed solves the issue in my case. 

I'm not entirely sure if this is a common issue for people as I did not find anyone else mentioning it in #sdk-help, but figured I would raise this issue if others are indeed facing the same problem.